Rock Paper Scissors
A very simple and basic game that almost every beginner in python can create. I wrote this program in 1 hour using the residual knowledge in my head. This program is still far from the final version, but I will try to improve it. There are many things you can change in it - interface, create animations, add more features, but right now I don't have enough knowledge for this, so now it's a very simple game.